# Conflicts:
#	src/nonebot_plugin_cs2match/config.py

+++ src/nonebot_plugin_cs2match/config.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2026 StarsetNight, XuanRikka
+# Copyright (c) 2023 StarsetNight
 # SPDX-License-Identifier: MIT
 
 from pydantic import BaseModel
@@ -17,5 +17,5 @@ class Config(BaseModel):
         ("pnl", 50, ["pnl"]),
     ]  # 赛事系列，优先级，匹配赛事名称（小写）
     client_timeout: int = 10
-    cache_ttl: int = 60
+    cache_ttl: float = 60
     cache_max_size: int = 64
